requirements.txt
Django==3.2
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2-binary==2.9.3
celery==5.2.3
django-celery-beat==2.2.1
requests==2.27.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
redis==3.5.3
django-redis-cache==3.0.0
pandas
gunicorn==20.1.0
whitenoise==6.0.0
drf-yasg
